modularize the serve function into lib
