+++ bin/serve.js
@@ -1,13 +1,11 @@
 var http = require('http')
 var getport = require('getport')
 var pump = require('pump')
+var route = require('../lib/serve.js')
 var abort = require('../lib/util/abort.js')
 var openDat = require('../lib/util/open-dat.js')
 var usage = require('../lib/util/usage.js')('serve.txt')
 var debug = require('debug')('dat-serve')
-var debugStream = require('debug-stream')
-
-var version = require('../package.json').version
 
 module.exports = {
   name: 'serve',
@@ -45,40 +43,15 @@ function startDatServer (args) {
         debug(req.method, req.url, req.connection.remoteAddress)
 
         if (req.method === 'GET') {
-          db.status(function (err, status) {
+          route.information(db, args, function (err, data) {
             if (err) abort(err, args)
-            db.listDatasets(function (err, datasets) {
-              if (err) abort(err, args)
-              res.setHeader('content-type', 'application/json')
-              var result = {
-                dat: true,
-                version: version,
-                status: status,
-                datasets: datasets
-              }
-              res.end(JSON.stringify(result))
-              return
-            })
+            res.setHeader('content-type', 'application/json')
+            res.end(data)
           })
         } else if (req.method === 'POST') {
-          var replicate = args.readonly ? db.push() : db.replicate()
-
-          var start = Date.now()
-          replicate.on('finish', function () {
-            debug('replication finished in', (Date.now() - start) + 'ms')
-          })
-
-          var loggers = {
-            out: debugStream(debugLogger('out')),
-            in: debugStream(debugLogger('in'))
-          }
-
-          if (!loggers.out.enabled) pump(req, replicate, res)
-          else pump(req, loggers.in(), replicate, loggers.out(), res)
-          return
+          return pump(req, route.replicate(db, args), res)
         } else {
-          // if not GET or POST
-          res.statusCode = 405
+          res.statuscode = 405
           res.end()
         }
       })
@@ -89,11 +62,5 @@ function startDatServer (args) {
 
       server.listen(port)
     })
-  }
-}
-
-function debugLogger (prefix) {
-  return function (buf) {
-    debug(prefix, {length: buf.length})
   }
 }
